requirements-test.txt

selenium==4.15.2
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-html==4.1.1
factory-boy==3.3.0
faker==20.1.0
coverage==7.3.2


django-extensions==3.2.3
plotly==5.17.0
pandas==2.1.3